cleaned up extra code to deploy
