+++ client/src/components/Login.js
@@ -1,5 +1,5 @@
 const postUserData = (name, code) => {
-  fetch('http://localhost:9000/user', {
+  fetch('http://localhost:3001/user', {
   method: 'POST',
   headers: {
     'Content-Type': 'application/json',
@@ -13,83 +13,3 @@ const postUserData = (name, code) => {
   .then(data => console.log(data))
   .catch(error => console.error('Error:', error));
 }
-
-
-// export default function loginForm() {
-//     return (
-//       <>
-//         <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">
-//           <div className="sm:mx-auto sm:w-full sm:max-w-sm">
-//             {/* uncomment this and add a logo */}
-//             {/* <img
-//               className="mx-auto h-10 w-auto"
-//               src="LOGO HERE"
-//               alt=" "
-//             /> */}
-//             <h2 className="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">
-//               Sign in
-//             </h2>
-//           </div>
-  
-//           <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
-//             <form className="space-y-6" action="#" method="POST">
-//               <div>
-//                 <label htmlFor="email" className="block text-sm font-medium leading-6 text-gray-900">
-//                   Email
-//                 </label>
-//                 <div className="mt-2">
-//                   <input
-//                     id="email"
-//                     name="email"
-//                     type="email"
-//                     autoComplete="email"
-//                     required
-//                     className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-//                   />
-//                 </div>
-//               </div>
-  
-//               <div>
-//                 <div className="flex items-center justify-between">
-//                   <label htmlFor="password" className="block text-sm font-medium leading-6 text-gray-900">
-//                     Password
-//                   </label>
-//                   <div className="text-sm">
-//                     <a href="#" className="font-semibold text-indigo-600 hover:text-indigo-500">
-//                       Forgot password?
-//                     </a>
-//                   </div>
-//                 </div>
-//                 <div className="mt-2">
-//                   <input
-//                     id="password"
-//                     name="password"
-//                     type="password"
-//                     autoComplete="current-password"
-//                     required
-//                     className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-//                   />
-//                 </div>
-//               </div>
-  
-//               <div>
-//                 <button
-//                   type="submit"
-//                   className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-//                 >
-//                   Enter if you dare
-//                 </button>
-//               </div>
-//             </form>
-  
-//             <p className="mt-10 text-center text-sm text-gray-500">
-//               Not signed up yet?{' '}
-//               <a href="#" className="font-semibold leading-6 text-indigo-600 hover:text-indigo-500">
-//                 Join the cool club now!
-//               </a>
-//             </p>
-//           </div>
-//         </div>
-//       </>
-//     )
-//   }
