+++ client/tailwind.config.js
@@ -6,7 +6,4 @@ module.exports = {
   theme: {
     extend: {},
   },
-  // plugins: [
-  //   require('@tailwindcss/forms'),
-  // ],
 }
