+++ server/model/User.js
@@ -1,46 +1,25 @@
-const { Schema, model } = require('mongoose');
-const bcrypt = require('bcrypt');
+const mongoose = require('mongoose');
+require('dotenv').config();
 
-const userSchema = new Schema(
-  {
-    username: {
-      type: String,
-      required: true,
-      unique: true,
-    },
-    email: {
-      type: String,
-      required: true,
-      unique: true,
-      match: [/.+@.+\..+/, 'Must use a valid email address'],
-    },
-    password: {
-      type: String,
-      required: true,
-    },
-  },
-  // set this to use virtual below
-  {
-    toJSON: {
-      virtuals: true,
-    },
+const userInfoSchema = new mongoose.Schema({ name: String, code: String })
+const User = mongoose.model('User', userInfoSchema);
+
+async function addUser(userInfo) {
+  try {
+    await mongoose.connect(process.env.MONGODB_URI);
+
+    var query = { name: userInfo.name } // find documents with matching name
+    console.log('n:', userInfo.name, 'c:', userInfo.code)
+    var update = { $set: {name:userInfo.name, code:userInfo.code} } // update if match found
+    var options = { new: true, upsert: true } // otherwise create new (upsert)
+
+    await User.findOneAndUpdate(query, update, options)
+
+    var users = await User.find({})
+    console.log(users)
+    
+  } finally {
+    mongoose.disconnect();
   }
-);
-
-// hash user password
-userSchema.pre('save', async function (next) {
-  if (this.isNew || this.isModified('password')) {
-    const saltRounds = 10;
-    this.password = await bcrypt.hash(this.password, saltRounds);
-  }
-  next();
-});
-
-// custom method to compare and validate password for logging in
-userSchema.methods.isCorrectPassword = async function (password) {
-  return bcrypt.compare(password, this.password);
-};
-
-const User = model('User', userSchema);
-
-module.exports = User;
+}
+exports.addUser = addUser
