+++ server/server.js
@@ -1,33 +1,27 @@
 const express = require('express');
+const db = require('./config/config.js');
+const { json } = require('express');
+const cors = require('cors');
+
 const { ApolloServer } = require('@apollo/server');
-// const { readFileSync } = require('fs');
-const { typeDefs, resolvers } = require('./schemas/index.js')
-// const {} = require('./resolver.js')
-// const resolvers = module.resolvers;
+const { typeDefs, resolvers } = require('./schemas/index.js');
 const authMiddleware = module.authMiddleware;
-// const gql = require('graphql-tag');
-const db = require('./config/config.js')
+const { addUser } = require('./model/User.js');
+const { expressMiddleware } = require('@apollo/server/express4');
 
 
 const app = express();
 const PORT = process.env.PORT || 3001;
 
-const uri = 'mongodb+srv://c00lwhip077:c00lwhip077@userinfo.njsmdfy.mongodb.net/zombieai?retryWrites=true&w=majority' // DELETE and use .ENV for KEY
+app.use(express.urlencoded({ extended: true }));
+app.use(express.json());
+  
+// if we're in production, serve client/build as static assets
+if (process.env.NODE_ENV === 'production') {
+  app.use(express.static(path.join(__dirname, '../client/build')));
+}
 
-// MONGOOSE CONNECTION
-// const mongoose = require('mongoose');
-// main().catch(err => console.log(err));
-
-// async function main() {
-//   await mongoose.connect(uri);
-// }
-
-// const UserSchema = new mongoose.Schema({
-//   user: String
-// });
-
-
-//Apollo server middleware and starting server
+// Apollo server middleware and starting server
 const server = new ApolloServer({
   typeDefs,
   resolvers, 
@@ -35,29 +29,14 @@ const server = new ApolloServer({
   persistedQueries: false, 
 });
 
-
-//
-
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-
-// const typeDefs = gql(
-//   readFileSync("schema.graphql", {
-//     encoding: "utf-8",
-//   })
-// )
-
-// const server = new ApolloServer({ typeDefs, resolvers })
-  
-// if we're in production, serve client/build as static assets
-if (process.env.NODE_ENV === 'production') {
-  app.use(express.static(path.join(__dirname, '../client/build')));
-}
-
-const startServer = async (typeDefs, resolvers) => {
+async function startServer () {
   await server.start();
-  // server.applyMiddleware({ app });
-};
+  app.use('/user', (req, res) => {
+    addUser(req.body)
+    res.sendFile((path.join(__dirname, '../client/build')));
+  })
+  app.use('/graphql', cors(), json(), expressMiddleware(server));
+}
 
 console.log(db)
